Envers and Validators tests(code cleanup )-JBQA-4985 and JBQA-4998
